+++ r-appservice.tf
@@ -1,3 +1,5 @@
+data "azurerm_client_config" "main" {}
+
 resource "azurerm_app_service" "app_service" {
   name                = local.app_service_name
   location            = var.location
@@ -50,6 +52,28 @@ resource "azurerm_app_service" "app_service" {
       name  = lookup(connection_string.value, "name", null)
       type  = lookup(connection_string.value, "type", null)
       value = lookup(connection_string.value, "value", null)
+    }
+  }
+
+  dynamic "auth_settings" {
+    for_each = var.auth_settings.enabled ? [var.auth_settings] : []
+    content {
+      enabled             = auth_settings.value.enabled
+      issuer              = format("https://sts.windows.net/%s/", data.azurerm_client_config.main.tenant_id)
+      token_store_enabled = var.auth_settings.token_store_enabled
+
+      unauthenticated_client_action = coalesce(var.auth_unauthenticated_client_action, "RedirectToLoginPage")
+      default_provider              = coalesce(var.auth_default_provider, "AzureActiveDirectory")
+
+      dynamic "active_directory" {
+        for_each = [auth_settings.value.active_directory]
+
+        content {
+          client_id         = active_directory.value.client_id
+          client_secret     = active_directory.value.client_secret
+          allowed_audiences = formatlist("https://%s", [format("%s.azurewebsites.net", local.app_service_name)])
+        }
+      }
     }
   }